Revert "Refactorings (#34)"
This reverts commit 12e8ba1d8965e47f65feb03dc957de8b77f98ac3.
