+++ src/App.tsx
@@ -3,7 +3,7 @@ import HomeScreen from "./components/HomeScreen/HomeScreen";
 import { Stack } from "@mui/material";
 import { createTheme, ThemeProvider } from "@mui/material";
 import { blueGrey, grey } from "@mui/material/colors";
-import ComicBook from "./components/ComicBook/ComicBook/ComicBook";
+import ComicBook from "./components/ComicBook/ComicBook";
 import { useEffect, useState } from "react";
 import comicService from "./services/comicService";
 import axios from "axios";
@@ -12,58 +12,8 @@ const theme = createTheme({
   palette: {
     primary: grey,
     secondary: blueGrey,
-    background: {
-      paper: grey[900],
-    },
-    text: {
-      primary: '#fff',
-      secondary: grey[300]
-    }
   },
-  components: {
-    MuiDialog: {
-      styleOverrides: {
-        paper: {
-          backgroundColor: grey[900],
-          color: '#fff'
-        }
-      }
-    },
-    MuiDialogTitle: {
-      styleOverrides: {
-        root: {
-          color: '#fff'
-        }
-      }
-    },
-    MuiDialogContent: {
-      styleOverrides: {
-        root: {
-          color: '#fff'
-        }
-      }
-    },
-    MuiTextField: {
-      styleOverrides: {
-        root: {
-          '& .MuiOutlinedInput-root': {
-            '& fieldset': {
-              borderColor: grey[700],
-            },
-            '&:hover fieldset': {
-              borderColor: grey[500],
-            },
-            '& input': {
-              color: '#fff'
-            }
-          },
-          '& .MuiInputLabel-root': {
-            color: grey[300]
-          }
-        }
-      }
-    }
-  }
+  typography: {},
 });
 
 function App() {
