+++ src/components/ComicBook/ComicPage.css
@@ -52,13 +52,8 @@
   padding-top: 3px;
   text-align: center;
   cursor: pointer;
-  filter: brightness(70%); /* Changed from 50% for better readability while maintaining visual hierarchy */
+  filter: brightness(50%);
   color: white;
-}
-
-.list-element-for-pages:hover {
-  filter: brightness(100%);
-  transition: filter 0.3s ease;
 }
 
 .focused {
@@ -100,7 +95,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  color: #FFFFFF; /* Changed from #c8c8c8 for better visibility */
+  color: #c8c8c8;
 }
 
 .answer-button {
