--- src/components/ComicBook/ComicPageView/ComicPageView.css
@@ -1,37 +1,0 @@
-.comicpage-in-modal {
-  margin: 0;
-  padding: 0;
-  max-width: 100vw;
-  max-height: 90vh;
-}
-
-.centered {
-  margin-top: auto;
-  margin-bottom: auto;
-  text-align: center;
-}
-
-.relative {
-  position: relative;
-}
-
-.fixed-top-left {
-  position: fixed;
-  top: 0;
-  left: 0;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.7);
-  z-index: 1000;
-}
-
-.comic-container {
-  width: 100%;
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 1rem;
-}
-
-.questions-container {
-  width: 100%;
-  max-width: 800px;
-  margin: 0 auto;
-}
