--- src/components/ComicBook/MathQuestions/MathQuestions.css
@@ -1,86 +1,0 @@
-.questionform {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.questions-table {
-  width: 100%;
-  border-collapse: separate;
-  border-spacing: 0 0.5rem;
-}
-
-.question-cell {
-  width: 80%;
-  text-align: left;
-  padding-left: 1rem;
-  color: #FFFFFF;
-}
-
-.answer-cell {
-  width: 20%;
-}
-
-.answer-input {
-  width: 100%;
-  height: 2.2rem;
-  padding: 0 0.5rem;
-  font-size: 1.2rem;
-  background-color: rgb(41, 41, 41);
-  border-radius: 4px;
-  border: 1px solid #ccc;
-  color: #FFFFFF;
-  box-sizing: border-box;
-  vertical-align: top;
-}
-
-.button-container {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  margin-top: 1rem;
-}
-
-.answer-button {
-  background-color: #374d46;
-  border: none;
-  color: white;
-  height: 2.2rem;
-  padding: 0 1rem;
-  text-align: center;
-  margin: 0;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  font-size: 1.2rem;
-  line-height: 2.2rem;
-  vertical-align: top;
-  width: 150px;
-}
-
-.answer-button:hover {
-  box-shadow: 0rem 0rem 3px rgb(255, 255, 255);
-}
-
-.wrong-answer {
-  background-color: #ff6b6b;
-  animation: flashRed 1s;
-}
-
-@keyframes flashRed {
-  0% { background-color: #374d46; }
-  15% { background-color: rgb(99, 0, 25); }
-  100% { background-color: #374d46; }
-}
-
-.right-answer {
-  background-color: #51cf66;
-  animation: flashGreen 1s;
-}
-
-@keyframes flashGreen {
-  0% { background-color: #374d46; }
-  15% { background-color: #00aa09; }
-  100% { background-color: #374d46; }
-}
