+++ src/components/HomeScreen/MenuComponent.test.tsx
@@ -1,9 +1,10 @@
-import { render, screen } from "@testing-library/react";
-import Navbar from "./Navbar";
-import { describe, test, expect } from "vitest";
+// import { render, screen } from "@testing-library/react";
+// import "@testing-library/jest-dom";
+// import MenuComponent from "./MenuComponent";
+// import { describe, test, expect } from "vitest";
 
 // TODO: fix file
-// describe("<Navbar />", () => {
+// describe("<MenuComponent />", () => {
 //   test("should render About author -button and show more -button", () => {
 //     // different screen sizes render different buttons,
 //     // but test does not recognize this
@@ -17,7 +18,7 @@ import { describe, test, expect } from "vitest";
 //     // setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
 //     // setOpenCreateAccount: React.Dispatch<React.SetStateAction<boolean>>;
 
-//     render(<Navbar setOpenAbout={setOpenAbout} />);
+//     render(<MenuComponent setOpenAbout={setOpenAbout} />);
 
 //     const element = screen.getByRole("button", {
 //       name: "about-author-big",
